+++ cmd/operations/sharpen.js
@@ -31,13 +31,15 @@ const queue = require('../../lib/queue')
 
 // Configure.
 const options = {
-  flat: {
+  m1: {
+    alias: 'flat',
     desc: 'The level of sharpening to apply to "flat" areas',
     defaultDescription: '1.0',
     nargs: 1,
     type: 'number'
   },
-  jagged: {
+  m2: {
+    alias: 'jagged',
     desc: 'The level of sharpening to apply to "jagged" areas',
     defaultDescription: '2.0',
     nargs: 1,
@@ -46,6 +48,24 @@ const options = {
   sigma: {
     desc: 'The sigma of the Gaussian mask',
     defaultDescription: '1 + radius / 2',
+    type: 'number'
+  },
+  x1: {
+    desc: 'The threshold between "flat" and "jagged" areas',
+    defaultDescription: '2.0',
+    nargs: 1,
+    type: 'number'
+  },
+  y2: {
+    desc: 'The maximum amount of brightening',
+    defaultDescription: '10.0',
+    nargs: 1,
+    type: 'number'
+  },
+  y3: {
+    desc: 'The maximum amount of darkening',
+    defaultDescription: '20.0',
+    nargs: 1,
     type: 'number'
   }
 }
@@ -63,7 +83,19 @@ const builder = (yargs) => {
 
 // Command handler.
 const handler = (args) => {
-  return queue.push(['sharpen', (sharp) => sharp.sharpen(args.sigma, args.flat, args.jagged)])
+  return queue.push(['sharpen', (sharp) => {
+    if (args.sigma || args.m1 || args.m2 || args.x1 || args.y2 || args.y3) {
+      return sharp.sharpen({
+        sigma: args.sigma,
+        m1: args.m1,
+        m2: args.m2,
+        x1: args.x1,
+        y2: args.y2,
+        y3: args.y3
+      })
+    }
+    return sharp.sharpen()
+  }])
 }
 
 // Exports.