+++ CHANGELOG.md
@@ -10,8 +10,12 @@
 * Added [`--id`](https://sharp.pixelplumbing.com/api-output#tile).
 * Added [`--lightness](https://sharp.pixelplumbing.com/api-operation#modulate).
 * Added [`--mozjpeg`](https://sharp.pixelplumbing.com/api-output#jpeg).
+* Added [`--resolutionUnit](https://sharp.pixelplumbing.com/api-output#tiff).
 * Added [`--timeout`](https://sharp.pixelplumbing.com/api-output#timeout).
+* Added [`--withoutReduction](https://sharp.pixelplumbing.com/api-resize#resize).
+* Added [`--x1`, `--y2`, `--y3`](https://sharp.pixelplumbing.com/api-operation#sharpen).
 * Drop Node.js 10 support, now requires Node.js >= 12.22.
+* Renamed `--reductionEffort` option to `--effort`.
 * Replaced `--squash` by [`--bitdepth`](https://sharp.pixelplumbing.com/api-output#tiff).
 * Updated default [`--compressionLevel`](http://sharp.pixelplumbing.com/en/stable/api-output/#png).
 * Updated dependencies.